Move legacy operators under the nn directory (#8858)
* Move legacy operators to nn/

* Move CuDNN operators to nn/

* Update CuDNN algoreg.

* fix amalgamation.

* fix a path of a header file.

* Fix the path for mkl header files.

* Fix compilation errors in mkl.

* Fix for the coding style.

* Fix the path of the header files.

* Fix headers in the cu files.

* Compile cu files.

* Update activation-inl.h

* Update activation-inl.h
