+++ amalgamation/mxnet_predict0.cc
@@ -64,16 +64,16 @@
 
 #include "src/operator/operator.cc"
 #include "src/operator/operator_util.cc"
-#include "src/operator/activation.cc"
-#include "src/operator/batch_norm.cc"
+#include "src/operator/nn/activation.cc"
+#include "src/operator/nn/batch_norm.cc"
 #include "src/operator/concat.cc"
-#include "src/operator/convolution.cc"
-#include "src/operator/deconvolution.cc"
-#include "src/operator/dropout.cc"
-#include "src/operator/fully_connected.cc"
+#include "src/operator/nn/convolution.cc"
+#include "src/operator/nn/deconvolution.cc"
+#include "src/operator/nn/dropout.cc"
+#include "src/operator/nn/fully_connected.cc"
 #include "src/operator/leaky_relu.cc"
-#include "src/operator/pooling.cc"
-#include "src/operator/softmax_activation.cc"
+#include "src/operator/nn/pooling.cc"
+#include "src/operator/nn/softmax_activation.cc"
 #include "src/operator/softmax_output.cc"
 #include "src/operator/tensor/elemwise_binary_broadcast_op_basic.cc"
 #include "src/operator/tensor/elemwise_binary_op.cc"
