+++ src/operator/convolution_v1.cu
@@ -27,7 +27,7 @@
 #include "./convolution_v1-inl.h"
 #include <vector>
 #if MXNET_USE_CUDNN == 1
-#include "./cudnn_convolution-inl.h"
+#include "./nn/cudnn/cudnn_convolution-inl.h"
 #endif  // MXNET_USE_CUDNN
 
 namespace mxnet {
