+++ src/operator/nn/activation.cc
@@ -24,11 +24,11 @@
  * \author Bing Xu
 */
 #include "./activation-inl.h"
-#include "./mshadow_op.h"
+#include "../mshadow_op.h"
 #if MXNET_USE_MKL2017 == 1
 #include <mkl_memory.h>
-#include "./mkl/mkl_memory-inl.h"
-#include "./mkl/mkl_relu-inl.h"
+#include "../mkl/mkl_memory-inl.h"
+#include "../mkl/mkl_relu-inl.h"
 #endif  // MXNET_USE_MKL2017
 
 namespace mxnet {
