+++ src/operator/nn/activation.cu
@@ -24,9 +24,9 @@
  * \author Bing Xu
 */
 #include "./activation-inl.h"
-#include "./mshadow_op.h"
+#include "../mshadow_op.h"
 #if MXNET_USE_CUDNN == 1
-#include "./cudnn_activation-inl.h"
+#include "./cudnn/cudnn_activation-inl.h"
 #endif
 
 namespace mxnet {
