+++ src/operator/nn/batch_norm-inl.h
@@ -23,8 +23,8 @@
  * \brief
  * \author Bing Xu, Chris Olivier
  */
-#ifndef MXNET_OPERATOR_BATCH_NORM_INL_H_
-#define MXNET_OPERATOR_BATCH_NORM_INL_H_
+#ifndef MXNET_OPERATOR_NN_BATCH_NORM_INL_H_
+#define MXNET_OPERATOR_NN_BATCH_NORM_INL_H_
 
 #include <dmlc/logging.h>
 #include <dmlc/parameter.h>
@@ -34,9 +34,9 @@
 #include <vector>
 #include <string>
 #include <utility>
-#include "./mshadow_op.h"
-#include "./operator_common.h"
-#include "mxnet_op.h"
+#include "../mshadow_op.h"
+#include "../operator_common.h"
+#include "../mxnet_op.h"
 
 #ifdef __GNUG__
 #pragma GCC diagnostic push
@@ -489,5 +489,5 @@ extern volatile bool disable_mkl;
 #pragma GCC diagnostic pop
 #endif
 
-#endif  // MXNET_OPERATOR_BATCH_NORM_INL_H_
+#endif  // MXNET_OPERATOR_NN_BATCH_NORM_INL_H_
