+++ src/operator/nn/convolution.cc
@@ -27,8 +27,8 @@
 #include "./convolution-inl.h"
 #if MXNET_USE_MKL2017 == 1
 #include <mkl_memory.h>
-#include "./mkl/mkl_memory-inl.h"
-#include "./mkl/mkl_convolution-inl.h"
+#include "../mkl/mkl_memory-inl.h"
+#include "../mkl/mkl_convolution-inl.h"
 #endif  // MXNET_USE_MKL2017
 #if MXNET_USE_NNPACK == 1
 #include "./nnpack/nnpack_convolution-inl.h"
