+++ src/operator/nn/cudnn/cudnn_activation-inl.h
@@ -24,11 +24,11 @@
  * \author Bing Xu
 */
 
-#ifndef MXNET_OPERATOR_CUDNN_ACTIVATION_INL_H_
-#define MXNET_OPERATOR_CUDNN_ACTIVATION_INL_H_
+#ifndef MXNET_OPERATOR_NN_CUDNN_CUDNN_ACTIVATION_INL_H_
+#define MXNET_OPERATOR_NN_CUDNN_CUDNN_ACTIVATION_INL_H_
 #include <algorithm>
 #include <vector>
-#include "./activation-inl.h"
+#include "../activation-inl.h"
 
 namespace mxnet {
 namespace op {
@@ -225,4 +225,4 @@ class CuDNNActivationOp : public Operator {
 };  // class CuDNNActivationOp
 }  // namespace op
 }  // namespace mxnet
-#endif  // MXNET_OPERATOR_CUDNN_ACTIVATION_INL_H_
+#endif  // MXNET_OPERATOR_NN_CUDNN_CUDNN_ACTIVATION_INL_H_
