+++ src/operator/nn/cudnn/cudnn_algoreg-inl.h
@@ -23,16 +23,16 @@
  * \brief
  * \author Bing Xu
  */
-#ifndef MXNET_OPERATOR_CUDNN_ALGOREG_INL_H_
-#define MXNET_OPERATOR_CUDNN_ALGOREG_INL_H_
+#ifndef MXNET_OPERATOR_NN_CUDNN_CUDNN_ALGOREG_INL_H_
+#define MXNET_OPERATOR_NN_CUDNN_CUDNN_ALGOREG_INL_H_
 
 #include <algorithm>
 #include <mutex>
 #include <string>
 #include <vector>
-#include "../common/cuda_utils.h"
-#include "./convolution-inl.h"
-#include "./deconvolution-inl.h"
+#include "../../../common/cuda_utils.h"
+#include "../convolution-inl.h"
+#include "../deconvolution-inl.h"
 namespace mxnet {
 namespace op {
 #if MXNET_USE_CUDNN == 1
@@ -175,4 +175,4 @@ typedef CuDNNAlgoReg<DeconvolutionParam> CuDNNDeconvAlgoReg;
 }  // namespace op
 }  // namespace mxnet
 
-#endif  // MXNET_OPERATOR_CUDNN_ALGOREG_INL_H_
+#endif  // MXNET_OPERATOR_NN_CUDNN_CUDNN_ALGOREG_INL_H_
