+++ src/operator/nn/cudnn/cudnn_pooling-inl.h
@@ -24,11 +24,11 @@
  * \author Bing Xu
 */
 
-#ifndef MXNET_OPERATOR_CUDNN_POOLING_INL_H_
-#define MXNET_OPERATOR_CUDNN_POOLING_INL_H_
+#ifndef MXNET_OPERATOR_NN_CUDNN_CUDNN_POOLING_INL_H_
+#define MXNET_OPERATOR_NN_CUDNN_CUDNN_POOLING_INL_H_
 #include <algorithm>
 #include <vector>
-#include "./pooling-inl.h"
+#include "../pooling-inl.h"
 
 namespace mxnet {
 namespace op {
@@ -311,5 +311,5 @@ class CuDNNPoolingOp : public Operator {
 }  // namespace op
 }  // namespace mxnet
 
-#endif  // MXNET_OPERATOR_CUDNN_POOLING_INL_H_
+#endif  // MXNET_OPERATOR_NN_CUDNN_CUDNN_POOLING_INL_H_
