+++ src/operator/nn/cudnn/cudnn_softmax_activation-inl.h
@@ -24,11 +24,11 @@
  * \author Bing Xu
 */
 
-#ifndef MXNET_OPERATOR_CUDNN_SOFTMAX_ACTIVATION_INL_H_
-#define MXNET_OPERATOR_CUDNN_SOFTMAX_ACTIVATION_INL_H_
+#ifndef MXNET_OPERATOR_NN_CUDNN_CUDNN_SOFTMAX_ACTIVATION_INL_H_
+#define MXNET_OPERATOR_NN_CUDNN_CUDNN_SOFTMAX_ACTIVATION_INL_H_
 #include <algorithm>
 #include <vector>
-#include "./softmax_activation-inl.h"
+#include "../softmax_activation-inl.h"
 
 namespace mxnet {
 namespace op {
@@ -181,4 +181,4 @@ class CuDNNSoftmaxActivationOp : public Operator {
 };  // class CuDNNSoftmaxActivationOp
 }  // namespace op
 }  // namespace mxnet
-#endif  // MXNET_OPERATOR_CUDNN_SOFTMAX_ACTIVATION_INL_H_
+#endif  // MXNET_OPERATOR_NN_CUDNN_CUDNN_SOFTMAX_ACTIVATION_INL_H_
