+++ src/operator/nn/deconvolution-inl.h
@@ -23,8 +23,8 @@
  * \brief
  * \author Wei Wu
 */
-#ifndef MXNET_OPERATOR_DECONVOLUTION_INL_H_
-#define MXNET_OPERATOR_DECONVOLUTION_INL_H_
+#ifndef MXNET_OPERATOR_NN_DECONVOLUTION_INL_H_
+#define MXNET_OPERATOR_NN_DECONVOLUTION_INL_H_
 
 #include <dmlc/logging.h>
 #include <dmlc/parameter.h>
@@ -34,8 +34,8 @@
 #include <vector>
 #include <string>
 #include <utility>
-#include "./operator_common.h"
-#include "./linalg.h"
+#include "../operator_common.h"
+#include "../linalg.h"
 
 
 namespace mxnet {
@@ -750,4 +750,4 @@ class DeconvolutionProp : public OperatorProperty {
 #endif  // DMLC_USE_CXX11
 }  // namespace op
 }  // namespace mxnet
-#endif  // MXNET_OPERATOR_DECONVOLUTION_INL_H_
+#endif  // MXNET_OPERATOR_NN_DECONVOLUTION_INL_H_
