+++ src/operator/nn/depthwise_convolution-inl.h
@@ -22,12 +22,12 @@
  * \brief CUDA depthwise convolution code
  * \author shuqian.qu@hobot.cc
 */
-#ifndef MXNET_OPERATOR_DEPTHWISE_CONVOLUTION_INL_H_
-#define MXNET_OPERATOR_DEPTHWISE_CONVOLUTION_INL_H_
+#ifndef MXNET_OPERATOR_NN_DEPTHWISE_CONVOLUTION_INL_H_
+#define MXNET_OPERATOR_NN_DEPTHWISE_CONVOLUTION_INL_H_
 #include <algorithm>
 #include <vector>
 #include "./convolution-inl.h"
-#include "../common/cuda_utils.h"
+#include "../../common/cuda_utils.h"
 
 #if MXNET_USE_CUDA
 #include <cub/cub.cuh>
@@ -350,4 +350,4 @@ void DepthwiseConvolutionOp<DType>::Backward(const OpContext &ctx,
 }  // namespace mxnet
 #endif
 
-#endif  // MXNET_OPERATOR_DEPTHWISE_CONVOLUTION_INL_H_
+#endif  // MXNET_OPERATOR_NN_DEPTHWISE_CONVOLUTION_INL_H_
