+++ src/operator/nn/depthwise_convolution_tf.cuh
@@ -26,8 +26,8 @@
 */
 #ifndef MXNET_OPERATOR_DEPTHWISE_CONVOLUTION_TF_CUH_
 #define MXNET_OPERATOR_DEPTHWISE_CONVOLUTION_TF_CUH_
-#include "../common/cuda_utils.h"
-#include "./mxnet_op.h"
+#include "../../common/cuda_utils.h"
+#include "../mxnet_op.h"
 
 namespace tf {
 namespace depthwise_conv {
