+++ src/operator/nn/fully_connected-inl.h
@@ -22,8 +22,8 @@
  * \file fully_connect_op-inl.h
  * \brief fully connect operator and symbol
 */
-#ifndef MXNET_OPERATOR_FULLY_CONNECTED_INL_H_
-#define MXNET_OPERATOR_FULLY_CONNECTED_INL_H_
+#ifndef MXNET_OPERATOR_NN_FULLY_CONNECTED_INL_H_
+#define MXNET_OPERATOR_NN_FULLY_CONNECTED_INL_H_
 
 #include <dmlc/logging.h>
 #include <dmlc/parameter.h>
@@ -32,9 +32,9 @@
 #include <vector>
 #include <string>
 #include <utility>
-#include "./operator_common.h"
-#include "./elemwise_op_common.h"
-#include "linalg.h"
+#include "../operator_common.h"
+#include "../elemwise_op_common.h"
+#include "../linalg.h"
 
 namespace mxnet {
 namespace op {
@@ -297,4 +297,4 @@ class FullyConnectedProp : public OperatorProperty {
 #endif
 }  // namespace op
 }  // namespace mxnet
-#endif  // MXNET_OPERATOR_FULLY_CONNECTED_INL_H_
+#endif  // MXNET_OPERATOR_NN_FULLY_CONNECTED_INL_H_
