+++ src/operator/nn/pooling-inl.h
@@ -24,8 +24,8 @@
  * \author Bing Xu, Jun Wu
 */
 
-#ifndef MXNET_OPERATOR_POOLING_INL_H_
-#define MXNET_OPERATOR_POOLING_INL_H_
+#ifndef MXNET_OPERATOR_NN_POOLING_INL_H_
+#define MXNET_OPERATOR_NN_POOLING_INL_H_
 
 #include <dmlc/logging.h>
 #include <dmlc/parameter.h>
@@ -35,8 +35,8 @@
 #include <vector>
 #include <string>
 #include <utility>
-#include "./operator_common.h"
-#include "./nn/pool.h"
+#include "../operator_common.h"
+#include "./pool.h"
 
 namespace mxnet {
 namespace op {
@@ -332,4 +332,4 @@ class PoolingProp : public OperatorProperty {
 }  // namespace op
 }  // namespace mxnet
 
-#endif  // MXNET_OPERATOR_POOLING_INL_H_
+#endif  // MXNET_OPERATOR_NN_POOLING_INL_H_
