+++ src/operator/nn/pooling.cc
@@ -26,8 +26,8 @@
 #include "./pooling-inl.h"
 #if MXNET_USE_MKL2017 == 1
 #include <mkl_memory.h>
-#include "./mkl/mkl_memory-inl.h"
-#include "./mkl/mkl_pooling-inl.h"
+#include "../mkl/mkl_memory-inl.h"
+#include "../mkl/mkl_pooling-inl.h"
 #endif  // MXNET_USE_MKL2017
 #if MXNET_USE_NNPACK == 1
 #include "./nnpack/nnpack_pooling-inl.h"
