+++ src/operator/nn/upsampling-inl.h
@@ -23,8 +23,8 @@
  * \brief
  * \author Bing Xu
 */
-#ifndef MXNET_OPERATOR_UPSAMPLING_INL_H_
-#define MXNET_OPERATOR_UPSAMPLING_INL_H_
+#ifndef MXNET_OPERATOR_NN_UPSAMPLING_INL_H_
+#define MXNET_OPERATOR_NN_UPSAMPLING_INL_H_
 
 #include <dmlc/logging.h>
 #include <dmlc/parameter.h>
@@ -34,7 +34,7 @@
 #include <vector>
 #include <string>
 #include <utility>
-#include "./operator_common.h"
+#include "../operator_common.h"
 
 namespace mxnet {
 namespace op {
@@ -332,4 +332,4 @@ class UpSamplingProp : public OperatorProperty {
 }  // namespace op
 }  // namespace mxnet
 
-#endif  // MXNET_OPERATOR_UPSAMPLING_INL_H_
+#endif  // MXNET_OPERATOR_NN_UPSAMPLING_INL_H_
